backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1
django-filter==23.3
psycopg2-binary==2.9.7
Pillow==10.0.1
reportlab==4.0.4
weasyprint==60.2
python-decouple==3.8
django-extensions==3.2.3
celery==5.3.4
redis==5.0.1
django-redis==5.4.0
django-storages==1.14.2
boto3==1.28.85
gunicorn==21.2.0
whitenoise==6.6.0
